Temporarily disable asynchronous loading of resources
The Genesis library is not safe for multi-threaded environments, particularly when reading multiple texture files. This feature will be disabled until I can recompile the library with the Microsoft multi-threaded runtime library. We also need to provide a 64-bit version of the Genesis library.

Alternatively, I could implement the geBitmap functions myself, but this would take significant time and it will not stop me from using it for more complex functionality like geMotion. It’s preferable to rely on the library for now, and in the future, we can develop our own implementation to reduce dependency on Genesis.lib