Add FGD property type for rotations (angles3d)
Considering that Genesis3D doesn't have a property key defined for entity rotations like Quake or HL does (angle, angles or mangle), we need to generalize it. This allow to rotate entities in Euler angles for a 3D vector.

Also in my testings, I notice that entity may not have a entity definition (like making a _SkyDome), so we need to be careful to check for nullptr.