Enable/disable flags based on brush/face type
In this case, it allows me to simulate radio buttons or mutually exclusive flags with checkboxes without doing too much refactoring.

This functionality is hardcoded by now because it takes times to extend this in a generic way for the GameConfigParser. Though, it would help to disable undefined behavior like Water and Lava being set at the same time.

Added setDisabledFlags for Flags widgets to disable certain checkboxes.

Only enabled for Genesis3D map format to keep the editor still able to open maps for other games.